Remove some extra code
